pyproject.toml
[project]
name = "llmcourse-hw3-turing"
version = "1.0.0"
description = "Multi-Agent Translation System with Error Analysis"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Tal Hibner", email = "tal@example.com"},
]

dependencies = [
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    "sentence-transformers>=2.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "torch>=2.0.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.4.0",
    "pylint>=2.17.0",
    "flake8>=6.0.0",
]

[project.scripts]
translation-pipeline = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.pylint.messages_control]
max-line-length = 88
disable = ["C0111", "C0103"]
